Fix event timestamps in live data
The TofEvent constructor epects a DateAndTime object, but we were
passing the pulse ID as a 64-bit uint.  The DateAndTime constructor
treated that as the number of nanoseconds since the epoch.  However, the
pulse ID is actually 2 32-bit fields concatenated together: seconds
since the epoch and nanosecond fraction.

The fix was to eplicitly (and properly) convert the pulse ID to a
DateAndTime object and pass that object to the TofEvent constructor.

Refs #6870