Do not load event data for adding data
Currently, the data inside the SANS2D files are composed to the event mode
data and the equivalent histogram data inside the same file. LoadNexus
loads the histogram data from the files. By, forcing loading nexus files
with LoadNexus, we will never load the event data.

This is not an issue, because this version of SANSadd2.py saves histogram
data. As so, the equivalent histogram will still be saved correctly.

re #6986