Ensure that the live listener background thread always exits
If the user canceles the live listener algorithm at just the right time,
it was possible for the background thread to be stuck in a loop waiting
for something from the foreground thread and thus the thread would never
exit.  This change adds a couple of checks to make sure the background
thread won't get stuck.

Refs #6661