Fix plot axes not updating on OS X.
The issue seems to be with the 'from' and 'to' boxes. When you enter a
value and click apply then the box does not lose focus like on the other
platforms so the internal value does not get updated and when the dialog
asks what the value is it returns the original and not the new one.

An old comment suggests that the fix used to cause a crash but it no
longer seems to...

Refs #7036