Make script usable for the current session of Mantid
In order to make the script usable as soon as it is downloade, it is not
enough to update the pythonscripts.directories as the last commit did, but
you have also to add to the python sys path at the current running session.

Running python code is not possible inside the Framework/ScriptRepository
level. So, a notification strategy was used in order to trigger a method
of MantidUI, which can execute the python code that is necessary to add
the new paths to the python sys.

re #7097