Re #6191. Add a new, much thinner class for data-detector mapping.
Usually, this information will be held in a workspace's collection of
ISpectrum objects and the existing (I)SpectraDetectorMap class will be
going away. However, there is occasionally a need to hold the mapping
separately from data. Examples are for the NearestNeighbours stuff and
if you need mapping for data that is not in a workspace (e.g. not yet
loaded in or workspace doesn't hold all spectra). This class will be
used in these circumstances but will be transient, not persistent (i.e.
the workspace will not have it as a member).