Re #6191. Use new SpectrumDetectorMapping in NearestNeighbours code.
I realise that there is a reduction in abstraction here as I haven't
replaced the ISpectraDetectorMap interface. It didn't seem sensible having
to reimplement the map iterators when this is the only place it is used.
What I have done is aliased the unordered_map that's passed in to
ISpectrumDetectorMapping.
The alternative I considered was placing SpectrumDetectorMapping in
Geometry, but I wanted the class to be immutable and that meant taking
in a MatrixWorkspace in the constructor.