Re #6191. Remove MatrixWorkspace::getWorkspaceIndexToDetectorIDMap.
It is redundant alongside the ISpectrum objects and was only being used
in tests and in one place in the ISIS SANS GUI (for which I've put in what
I believe to be equivalent code, though there's no test to verify it (too
much logic in the GUI layer).
This also resolves part of #6198.