Add the transmission information to the reduced workspace
Due to some limitations imposed by the way the Singleton is used, there is
no way to get the transmission workspaces after the reduction is finished
from the user interface, except using the attempt to guess the name of the
transmission workspaces (what we should not do). So, I propose here keeping
the information about the workspaces of the transmission at the SampleLog
of the Reduced workspace. This information will be them available whenever
is necessary. It would even allow us to use this information inside the SaveCan
algorithm itself. But, I think it would seem a litle bit of magic. So, i
prefer to use it outside.

re #6929