Append newline at QtSignalChannel level rather than MessageDisplay
The Poco log stream emits the log messages when it receives a newline but
doesn't send along newlines. However, printing from Python scripts does
produce newline characters so in order not to double-count them it makes
most sense to simply emit a newline in the SignalChannel and let
everything listening to it simply insert the exact text received.

Refs #7152