RE 7091 bias data in splatterplot
Initial version.  Now sorts lists of boxes in
decreasing order and takes points in order from the
ordered list of boxes.  If a large number of points
are requested, the number of points actually used
may be much smaller.  The code steps along the list
trying to get the AVERAGE number of points per box,
from each non-empty box, with a minimum of of one
point per box.

refs #7901