RE 7091 bias data in splatterplot
Removed a couple of variables intended for an alternate
biasing strategy that won't be used.  Put cout statements
in if (VERBOSE) structure that was originally present in the
code.  Tried an alternate biasing strategy that used points
from the boxes above the threshold, based on the number of points
in the box.  For example, if the number of points requested was
equal to 1/2 the total number of points in boxes above the
threshold, then 1/2 of the points from each box would be used.
This alternate strategy emphasized bright peaks much more
than the weaker peaks and required many more points to see the
full lattice.  Also, it was more difficult to get a good
view of scattering between the main Bragg peaks in TOPAZ 7424
or of the diffuse scattering in TOPAZ 7238, with the alternate
strategy.

refs #7091