RE 7091_bias data in splatterplot
Added some additional comments in the code on the strategy being used.
Basically, the average number of points to use per box is calculated
as the number of points in the boxes above the threshold, divided by
the number of boxes above the threshold.  The list of boxes is then
processed in decreasing order, taking at least one point from each
box, up to max(average_points_per_box,number_in_box), until all boxes
have been processed or the requested number of points has been used.
This strategy seems to work quite well and both high and low level
features in a run such as TOPAZ 7424 can be seen.

refs #7901