Refs #7212. Forcing display update to set plot axes.
Due to the mechanism used to generate the SliceViewer views from the VSI
MultiSlice view, when a rebinner was used to reorient the axes, a mismatch
arose between the cached dimensions and the workspace dimensions. I've forced
an axis update to ensure that things are commensurate. The change is done in a
function that is ONLY called by the shift-click mechanism in the MultiSlice
view. This limits the side effects to this change.
(cherry picked from commit 1056e0bac06163bd32d0fa1732ad469344012e25)