Refs #7212. Fixing issue with integrated "time" dimension in VSI.
This fix was necessary to handle an annoying issue with the SEQ_MDEW.nxs file.
When the SliceViews were generated by the shift-click mechanism, they kept
coming out blank. I traced this due to a bad time value being handed to the
SliceViewer. Datasets with an integrated time axis need to be treated slightly
more carefully.