RE 7140_MantidEV_load_DetCal
Initial version.  Added controls for entering .DetCal
file names for two .DetCal files.  (The second file is
only needed for the second panel on SNAP.)
Added code to enable/disable the new controls properly.
Added code to save/restore the new controls' state properly.
Added code to record/save/restore names of most recently
used calibration files, so the file browsers for the
.DetCal files can start in that directory.
Factored out some common code for finding the directory
in which to start, based on a file name, placing it in
the method getFilePath().

refs #7140