RE 7140_MantidEV_load_DetCal
Place new controls for loading .DetCal file before the
controls for Max |Q| and selecting the Lorentz correction
so the order is consistent with the order in which the
algorithms are applied.
Also resetting the default state using the option on the
File menu will reset the .DetCal file names to blank.

refs #7140