RE 7244_MantidEV_misses_Monoclinic_C
Replaced ScalarUtils::GetSignRelatedUBs() with
ScalarUtils::GetRelatedUBs().  The new method not only tries
negating pairs of edges, but also forms a set of permutations
of the edges that both preserve handedness and could still
form Niggli cells to within experimental error.  With this
fix MantidEV now finds the Monoclinc C cell, so this fixes
the problem.
The test code still needs to be updated for this new
method.

refs #7244