RE 7278_Bounds_issue_in_vtkSplatterPlotFactory
The points, signals and number of points per vtk cell is now
recorded in vectors while going through the list of boxes
above the specified threshold.  With that information, vtk
data structures are initialized to hold exactly the number
of points and cells that will be used.  The points, signals
and number of point per cell info is then copied into the
right-sized vtk data structures.
This fixes the problem with calculating proper bounds for
the splatter plot data and consequently fixes the problem
with data disappearing from the splatter plot view when
pick mode was created.
NOTE: The original splatter plot code also created vtk
data structures for the signals and vtk cells that were
too large, but since picking was not done, the problem
was not discovered.  That original problem was carried
forward in the code when the strategy for including
points was revised to produce a useful view.

refs #7278