RE 7248_MantidEV_show_distance_between_last_points_picked
The getUB() method of MantidEVWorker now includes a parameter
indicating whether or not the UB should be converted to
lab coordinates.  If lab_coords is true, the goniometer
rotation matrix is obtained from the first peak in the
peaks workspace and is multiplied times UB to get a UB
in lab_coordinates.

refs #7248