refs #6667 Maps reduction works
But there is subtle difference in reduction results

the difference does not have big physical sense, as the relative values of differences are small, but still unclear why it is present at all.

The CheckWorkspacesMatch got possibility to analyze relative error together with the absolute error to understand the difference better
LoadMask.cpp in the ISIS part has been changed to report loaded masks on debug level rather then notice as the messages in the log complicate understanding of what is going on.