Allow full data path to be given with folders having dots
The issue of the system test was related that the ubuntu and mac os path name
had a folder with a dot. And the input was expecting only a file_name'.'extension.

Now, if it finds more than one dot, it will get only the last one to split
the file - extension.

re #7221