Add HDFDescriptor class that will describe HDF files.
The checks for the headers at the start of HDF4 & HDF5 files have been
put in a static isHDF member to encapsulate the details about
checking for a HDF file and make it reusable.
It will not retain an open file handle to the file so does not
inherit from the standard FileDescriptor.
Refs #7263