Changed checkSizeCompatiblity to std::string to aid error usefulness
BinaryOperation::checkSizeCompatiblity and overridden versions in child
classes have been changed to return std::string rather than bool so that a
useful error message can be passed back down the call stack before finally
being used in throwing std::invalid_argument in BinaryOperation::checkCompatibility.
A return of true is now a return of "" (empty string). A return of false is
now a string of length 1 or more containing an error about why the
algorithm couldn't complete. Functions that used the bool return before
have been changed to use checkSizeCompatiblity(lhs,rhs).empty as that is
the equivelent of the old boolean return.
Removed a load of commented-out code from Multiply.cpp.
Indented WeightedMean.cpp and WeightedMean.h as per the guidelines.

Refs #4489