Send to options working more sensibly
Changes have been made to the tree in View->Preferences->Mantid->Send To
so that it opperates more sensibly, in specific:
Checkboxes in the tree now switch visibility in the menu on or off.
the "visible" option in the edit dialog has been removed in favor of using
the tree checkboxes.
selection to edit is now made by selecting the item rather then clicking
the checkbox.
functionality of ConfigDialog::createCheckedTreeItem now makes sence based
on the parameters, passing true as the second paramter will now produce a
checked item. This will have broken some other functions that use, it, but
it'll only require that i negate their current value they pass.

Refs #4495