Provide a new interface for the calibration framework
This module will be the interface for the calibration framework.
It provides internal documentation that will be make available on #7387.

It also provides a new interface with different arguments for the calibration,
in order to not require any special knowlegde of internal objects to interact
with the calibration.

So, now, it is possible to calibrate an instrument just providing the workspace,
the name of the tubes to calibrate and the expected positions.

re #7387