refs #7372. Give the GUI a proper name.
This is a bit hackey, because Max has lost the original *.ui file, and clearly that would be the best place to fix it. However, it's going to be a semi-major operation to reverse-engineer the ui file out of this generated file. We do plan to do it at some point.