Re #7462. Add a live button to the MWRunFiles widget.
By default the button is hidden and disabled, so existing uses of this
widget will be entirely unaffected.

A property has been added so that the button can be set to be always
hidden and disabled, only shown if able to connect to the default
instrument or always shown (but only enabled if able to connect).
Selecting the live button will disable the file text box/browse button
& vice versa.