Use the detectors from baseInstrument
It nows use the position from the baseInstrument which is free from any
previous calibration being applied to the detectors.

The only issue that it introduces is that you can not calibrate instruments
that was moved or rotated before the acquisition of the data. Because,
the baseInstrument does not consider this. Besides, any previous movement
for the detector will be reset when applying the output of the calibration.

But, we consider these limitations better than having to switch off loading
the calibration values through comment the region det-pos-source.

re #7491