Change Quadratic to Polynomial n2
Changint the fit curve from quadratic to polynomial with n=2 produces
exactly the same result. This was seen by saving the calibtable to a file

{{{
tb = mtd['CalibTable']
tofile = open('/tmp/quadraticcalibrated.txt','w')
for row in tb:
	print >> tofile, row['Detector ID'], row['Detector Position']
tofile.close()
}}}

and comparing the files for the quadratic and for the polynomial n=2, and
both files are identical.

re #7501