Add separate facilties file for use in unit tests.
The TEST facility has been removed from the main Facilities.xml file and
into its own test one. The affected unit tests now load the test file if
and restore the original at the end. The FacilityHelper file defines
a simple struct to does this with an object in a scoped manner.

The test file is in the IDFs_for_UnitTesting sub-directory so it does not
get included in the packages.
Refs #7487