refs #7013. Expose PeaksViewer via SliceView
Refactored the SliceViewer so that the aspects creating the peaks presenters and displaying them were extracted to a separate public method. Then exposed these methods to python, and provided unit tests for the functionality. Unit tests have been set up such that if a sleep is introduced, you should see the marker sitting over a fake Bragg peak.