Re #5896. Ensure signal is emitted if state changed programatically.
Switch to a signal that is emitted whenever the state of the button
changes, not just when it is manually clicked.
This ensures that items are properly disabled if the live button starts
out checked. It appears to have no negative effect on the Step Scan GUI
(which is the other place listening to that signal).