Re #7524. Use more sensible default values.
For the vector of pointers to event lists that is used when adding the
events, make the default NULL instead of the first event list so that
detector ids that aren't in the workspace but might exist don't point
to the wrong thing. NULL-ness is already checked for at the point of
use.

Also, add a check on the vector that comes back from
MatrixWorkspace::getDetectorIDToWorkspaceIndexVector for the workspace
index being out of range because it has the very large default. In that
case no attempt is made to assign an event list pointer at it remains
at the default (NULL).