Re #5896. Use user-supplied workspace name for output.
Also put in code to get around the problem that only the first live
chunk will have an entry in the EnergyRequest log. For subsequent
chunks we have to copy it from the processed workspace.

This code will need cleaning up later. Probably some will be pushed up
to a higher layer and perhaps the output workspace can go in the
GeneralSettings so that we don't have to go fishing for it.