Correct the definition of the proxy required from HTTPClientSession
It was found that the error was that the HTTPClientSession required
from setProxy only the host, passing the 'http::<host>' caused it to fail.

Now, the host and the port are being queried indirectly through Poco::URI,
and this solves the issue. But, we considered that it was a good practice
to test the connection, given that it will be tested only once in the
startup of mantid and will be available afterwards.

re #7571