Xtof commit of mantid4POLDI algorithms
Update of the algorithms for Mantid. The different ipp files are now manage by the ProjectManager.
A bunch of ipp files, from 2009 to 2013 are delivered.
Some others algo are updated.
Following is the list of the modified files, with linked comments

update of the python API access, no more hard links
	modified:   Code/Mantid/Framework/PythonInterface/plugins/algorithms/LoadSINQ.py
	modified:   Code/Mantid/Framework/PythonInterface/plugins/algorithms/LoadSINQFile.py

gestion of the different ipp files
	modified:   Code/Mantid/Framework/PythonInterface/plugins/algorithms/PoldiProjectAddDir.py
	modified:   Code/Mantid/Framework/PythonInterface/plugins/algorithms/PoldiProjectAddFile.py
	modified:   Code/Mantid/Framework/PythonInterface/plugins/algorithms/PoldiProjectRun.py

useless algo for users, as to be used only by Mark or POLDI scientist
	deleted:    Code/Mantid/Framework/PythonInterface/plugins/algorithms/poldigroup.py

a non used algo doesn't need to be compiled
	modified:   Code/Mantid/Framework/SINQ/CMakeLists.txt

the constante are now properly declared
	modified:   Code/Mantid/Framework/SINQ/inc/MantidSINQ/PoldiAutoCorrelation5.h

nothing I think, just touch
	modified:   Code/Mantid/Framework/SINQ/src/LoadFlexiNexus.cpp

the starting time of the experiment is now used to choose the proper IDF
	modified:   Code/Mantid/Framework/SINQ/src/PoldiLoadLog.cpp

reset() is no more needed, it is automaticaly done (to be check)
	modified:   Code/Mantid/Framework/SINQ/src/SINQHMListener.cpp

grouping file moved in the proper dir and named properly
	renamed:    Code/Mantid/instrument/nexusdictionaries/poldigroup.xml -> Code/Mantid/instrument/Grouping/POLDI_Gro
uping_800to400.xml

IDF and parameters for years 2009, 2010, 2011, 2012, 2013
	new file:   Code/Mantid/instrument/POLDI_Definition_ipp10.xml
	new file:   Code/Mantid/instrument/POLDI_Definition_ipp11.xml
	new file:   Code/Mantid/instrument/POLDI_Definition_ipp12.xml
	new file:   Code/Mantid/instrument/POLDI_Definition_ipp13.xml
	new file:   Code/Mantid/instrument/POLDI_Definition_ipp9.xml
	new file:   Code/Mantid/instrument/POLDI_Parameters_ipp10.xml
	new file:   Code/Mantid/instrument/POLDI_Parameters_ipp11.xml
	new file:   Code/Mantid/instrument/POLDI_Parameters_ipp12.xml
	new file:   Code/Mantid/instrument/POLDI_Parameters_ipp13.xml
	new file:   Code/Mantid/instrument/POLDI_Parameters_ipp9.xml

Refs #100