Remove mutex lock from WorkspaceGroup::createInfoNode
This is called by the base Workspace::addInfoNodeTo method when the object
is a WorkspaceGroup. The base method locks the workspace RWLock but then
for a group the overloaded createInfoNode was then also locking a mutex
that causes a dead-lock while both locks wait for each other to release.
Refs #7677
(cherry picked from commit 59477d7a6745832606b84ef724b707ae3022de49)