Fix a minor bug in naming of group members in ChopData.
The ADS tries to add the members of a group to the ADS if they don't have
a name. It's naming scheme is different to what ChopData does and
so you effectively end up with an object having two names. The dock
then reports workspaces that don't exist and you can't delete them.
Refs #7677