Fixed some bugs in AlgorithmFactory
There were a couple of rather major bugs which it's a wonder hadn't been found until now.

The first was that decodeName() could go out the range of the string and return garbage if there wasn't a bar cahracter in the supplied string. A check has been added which will throw an exception if one isn't found.

The second was that getCategories() didn't work at all as the loop wasn't executing due to the variable "it_end" used in the condition not being initialised correctly. It's now initialised correctly to categoryMap.end().

Refs #3344