Fit options for SANS Reduction transmission
this commit provides a more flexible way to deal with the fit for the
transmission for SANS reduction.

Now, you can set the option for the transmission and select a different fit
option for the can, or for the sample.

This commit does not change anything. It just rearranges things, to prepare
the path to the introduction of more fit options and to use different fit
options for each transmission.

re #7733