Make FilePropertyWidget use default extension.
Before these changes it was using default extension for save dialogs only.
The point here is that code for save and load dialogs is doing almost the
same thing and should probably be moved to a single method. Additionally,
for some reason load dialogs are using ".ext (*.ext)" approach, while
save dialogs are displaying "*.ext" straight away. Somebody should probably
take a look at that as a maintenance.

Refs #7772