refs #7664. Update the UI components for new behaviour.
This hooks the UI into the new algorithms. The menu will now use HasUB to determine if the workspace has any UB matrixes, and update the clear options appropriately, this includes handling of workspace types that are not compatible with ExperimentInfo, such as table workspace. The Clear UB option is disabled until HasUB reports otherwise. The Clear UB option should execute the ClearUB algorithm.