Change the way to get the limit positions from workspace
Before, it tried to use the getBoundingBox from the instrument. But, if
there were no sample attached to the workspace, a segmentation fault was
triggered.

Now, it will try to get the limits of the position from the detectors, by
querying the first available detector and the last one, and from their position
guess the region defined by the instrument.

re #7776