refs #7535. TextAxis to return min and max without throwing
This is the route cause of the issue with the SliceViewer described by this ticket. If the Text axis stores first and last entries that are not numeric, then a boost lexical cast exception is generated upon getMin and getMax. The fix is to treat this methods as though they have throw() modifiers. Firstly, attempt to cast the first and last values, but if this fails, then fall-back to default numeric values. Unit tests added to verify this new behaviour. Also introduced fix in MatrixWorkspace, MWDimension, whereby, the getMin, getMax were not calling the equivalent methods on the wrapped Axis type.