Re #7785. SNS listener throws a NotYet exception when appropriate.
In particular, it throws it if data was requested from the start of the
run but the instrument is not in a run. This allows the user to cancel
the algorithm or wait until a run starts, at which point the data will
flow.
It will also throw if the workspace has not been initialized, though
not immediately as experience shows that it can sometimes (SEQUOIA for
example) take some time for the instrument to be loaded.