Finish converting remote algorithms to new API
Finished converting the remote algorithms to the new API.
Note: The query and submit algorithms haven't been tested yet because
they must be tested against Fermi and that can only be done from the
analysis machines.  Checking the code in now so that I can pull it down
to an analysis machine and try it.

Refs #7769