refs #7873. Fix ConvertUnits and add unit tests to confirm.
The fix is to check whether the consumer seems to intend that the input workspace is to replace the output workspace, or not. Clone the workspace if not. Note that I also had to make a minor change to the instrument param file path, because this wasn't working on my machine at all.