Set Precount option LoadEventNexus to true by default.
It prevents the vectors from being over allocated by push_back and
means the deletion on Windows is then much quicker. The memory
fragmentation should also decrease as the allocation to the correct
size is done once per vector.

The performance tests will have to be monitored to check on the overall
effect on speed when combined with tcmalloc on linux. On windows, at least,
the loading is faster.
Refs #7780