Re #7950. DGS Reduction GUI now starts up successfully.
Note on the changes in util.py: the old QString.toDouble/Int methods
returned a tuple with the 2nd value being a boolean indicating success.
The python conversion operators throw a ValueError for invalid input.
The QString method behaviour was to set the returned value to zero on
failure - I've retained that behaviour.