Re #7950. Set PyQt API version for stand-alone running.
If running in Mantid, this setting is done within the C++ code. I'm adding
it here to make sure things work if running outside of MantidPlot (which
doesn't work at the moment, but could cause confusion later if that's fixed).
Note that this has to be done before the first import of PyQt.