Re #7822. Keep the last log entries instead of clearing.
The previous way presented problems with unchanging or
slowly varying logs as a chunk could be left with an empty
log that the analysis relied on having.
The particular problem one was 'EnergyRequest', which was
only appearing in the first chunk.
I think that we can count on the logs coming in in time order, so the
fact that the underlying methods don't check or sort should not be a
problem.