Made the title bar captions more descriptive
The title bar cations are no all prepended with the workspace name.

They are also no longer truncated to the first hyphen. If there was a reason for that please tell me, but i couldn't see any reason why they were, especailly as the logs use underscores as spaces which are then converted to hyphens (for a reason that is explained in a comment) so it previously meant that multi-worded log names were truncated.

To Tester:

Make sure that the title bar captions show the workspace name, log name and a number which should be a fallback inidivitual identifier.

Be aware that this ticket was done while #7968 was awaiting merging into master thus the bug fixed there reared its head; it should only cause debug assertions on windows but this should be kept in mind.

Refs #8031