Re #6460. Only apply 'CheckType' option to event workspaces.
This algorithm property is only applicable to event workspaces (as the
property documentation states) but was being used for MD & Peaks
workspaces. However, the outcome if you pass in one of these workspace
types along with something else is a crash if the CheckType property is
false.

Therefore, always enforce a match for these types.