Refs #7429. After further review, permanently removing bounds checks.
Due to the restriction of weak ordering on the sorted boxes and that all the
boxes save the same signal strength, the sorting order of the boxes is
arbitrary. Therefore, the associated events (one per box) can be at any
location, so bounds checking the final VTK dataset is impossible.