Refs #7970 - Add alternate syntax for cuboid.
Added unit tests to cover the default axis and centre fields.

It is not possible to enforce an "either/or" syntax rule in the schema
file without wrapping each syntax in its own pair of tags.  We have to
leave that job to the ShapeFactory when it parses the XML.

I would have preferred to keep CuboidCorners and the new parseCuboid
function in an anonymous namespace, but had to list them in the header
file to make use of the other member functions of the factory.