refs #8125. Handle cycles for each instrument.
Allow switching between cycles for a single instrument. Display cycles to choose from. Update cycles when instrument is changed. Default to latest cycle and load run names.

Also set it up so that it extracts the run number from the workspaces as part of the transfer process so that only run number rather than workspace names appear in the table.