New version of LaodAscii written
The new verison of LoadAscii was going to be based on the original, and it has been, but not as much as was initially intended, and it ended up a full rewrite.

The Loader will load the form of files saved by SaveAscii version 2, but not version 1. The format has changed that much.

It is slgihtly flexible in that it will load files that SaveAscii wouldn't wirte, but htis only extends as far as it'll accept blank lines as equivelent to spectra IDs, will allow 2 collumns minimum (implying that E and Dx are 0), and will ingnore lines staritng wiht a comment character. However, it is still strict enough that it will expect any standards set at the start of the file to be continued thoughout. This means number of bins per spectra, number of columns per bin, inclusion or omittance of spectra IDs, comment characters and delimiters.

Refs #7732