Re #8159. Make sure the filename property's name is set correctly.
In the case of LoadEventPreNexus is is "EventFilename". It seems that for
everything else is is "Filename" because the variable that holds this is
empty by the time the algorithm comes to run (though the loader is
selected correctly).
I've factored out the method that sets the variable and called it again
from within a method that exec() calls.