Removed a redundant method, and corrected some things.
isAscii() was originally included as I copied LoadAscii v1 originally, and as it's not actually used internally i've removed it form v2. v1's isAscii is only used once (and by another class), and LoadAscii (the v1 class) is specifically used then, so it's safe to remove it from v2. In theory, that call to v1 should probably be using a file descriptor instead.

The spacers 2d array now has it's bounds corrected for both the new item, and reducing the second dimension as much wasn't used.

The Filename property has had an error in it's description corrected.

Refs #7732