Re #8259. Delay showing of interpreter/IPython widget.
This widget was being created 'early' in the initialization phase, with a
justification given that it had to exist when the settings are read (which
hold its previous location and whether it was visible) and
QMainWindow::restoreState() called. In fact, this is not a requirement as
there's a method QMainWindow::restoreDockWidget() that can be called later
if a widget is created after restoreState() is called.

I've switched to using that and on linux this makes a floating IPython
window appear at moreorless the same time as the main application window.