Refs #7732 LoadAscii2 now allows - and + at start of line
A slight mistake whenre i was only checking for digits at the start of valid lines. The Algorithm will now allow + and - if supplied to determine positive and negative.
lineNo is now a member variable in order to fix line number accuracy after the changes with processHeader made them incorrect.