re #8395: creation of Larmor Instrument
This commit introduces the Larmor instrument that is required to allow
reducing Larmor data. It was not possible to check with the user the
system coordinates that he wants to use and other things.

But, this commit is necessary to allow the user start interacting with
the reduction to allow us to address all the particularities of the
reduction for LARMOR.