Refs #8475 Fixing a dimension
I'd resized the hight to test that the other widgets resized too, but i'd forgotten to put it back.

It's worth noting that the old gui had a width of 1300 which wasn't suitable as that's wider that the most common resolution in use, 1024x768.

The default width is now 1000, the height remains 400.