Refs #8745 Replicated the original functionality
UI has had another widget added that I'd missed before because it was invisible

Original functionality has been ported over with a couple of fixes.

Save Workspaces has been disabled for now as all that window did was run SaveAscii with default values on a workspace already in mantid, so the dialog was rather redundant as the user could run SaveAscii (or a more preffered Save algorithm) from within mantid.