refs #8372. Factor out coAdd
I'm leaving coAdd in the script for now, although the new convert_to_wavelength code will actually do the job of this code. coAdd actually does a lot of different and bizarre things in a single method, including StartLiveData, which is not yet converted by convert_to_wavelength as the current implementation of coAdd requires 'magic' decision making depending upon the type and format of the 'run' argument as well as the current state of the ADS etc.