Refs #8509 processheader() now checks more thoroughly
processheader() used to get called after checking the base number of collumns, which was what was cusing the error.

processheader() is now called before setting the base number of collumns and will set the base number of columns as it checks, but only if the user doesn't specifiy a number of columns to skip. processheader() will check for a number of valid lines with the same number of columns (between 2 and 4 inclusive), skipping over and commented or blank lines, and resetting the matching counter when it hits an invalid line of any sort.